% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_walk.R
\name{fit_walk}
\alias{fit_walk}
\title{Estimate the parameters of a random walk (mu and sigma)}
\usage{
fit_walk(positions)
}
\arguments{
\item{positions}{a sequence of positions}
}
\value{
fitted walk estimated parameters
}
\description{
Estimate the parameters of a random walk (mu and sigma)
}
\examples{
fit_walk(walk)
}
