% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_walk.R
\name{plot_walk}
\alias{plot_walk}
\title{Plots a random walk}
\usage{
plot_walk(walk)
}
\arguments{
\item{walk}{a simulated random walk}
}
\value{
plot a random walk over 100 time steps.
}
\description{
Plots a random walk
}
\examples{
plot_walk(walk)
plot_walk(simulated_walk)
}
