% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_walk.R
\name{simulate_walk}
\alias{simulate_walk}
\title{Simulate a random walk}
\usage{
simulate_walk(t, mu, sigma)
}
\arguments{
\item{t}{time steps}

\item{mu}{Step lengths drawn from Normal(mu, sigma)}

\item{sigma}{Step lengths drawn from Normal(mu, sigma)}
}
\value{
a simulated random walk
}
\description{
Simulate a random walk
}
\examples{
simulate_walk(t=100, mu=0.5,sigma=2)
simulate_walk(t=200, mu=0.8, sigma=2.5)
}
